Extract _forms.py from _legacy.py
